requirements/main.txt
# Main requirements for the library
numpy>=1.24.2
scipy>=1.10.1
pandas>=1.5.3
tqdm>=4.65.0
wandb>=0.15.3
pyarrow>=13.0.0
pandas
wandb>=0.15.3
matplotlib>=3.6.1
vit_pytorch>=1.6.5
lightning>=2.1.3